% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function01.R
\name{make_dna}
\alias{make_dna}
\title{Make a random dna sequence of chosen length}
\usage{
make_dna(number)
}
\arguments{
\item{number}{integer. Defines the length of the sequence}
}
\value{
"ATGTTCGCAAAAA"
}
\description{
Make a random dna sequence of chosen length
}
